Ajout de la fonction downloadImage dans le modele Pokemon pour s'accorder avec le MVC
