+++ iPokedex/view/DetailViewController.swift
@@ -62,7 +62,7 @@ class DetailViewController: UIViewController {
                     self.defenseLabel.text = "Defense : \(pokemonDetail.stats[3].base_stat)"
                     self.baseSpeedLabel.text = "Speed : \(pokemonDetail.stats[0].base_stat)"
                     
-                    self.downloadImage(urlString: pokemonDetail.sprites.front_default!, imageView: self.pokemonPicture)
+                    self.currentPokemon?.downloadImage(urlString: pokemonDetail.sprites.front_default!, imageView: self.pokemonPicture)
                     
                     self.activityIndicator.stopAnimating()
                     self.activityIndicator.isHidden = true
@@ -89,26 +89,7 @@ class DetailViewController: UIViewController {
         //navigationController?.pushViewController(<#T##viewController: UIViewController##UIViewController#>, animated: <#T##Bool#>)
     }
 
-    func downloadImage(urlString: String, imageView: UIImageView){
-        imageView.contentMode = UIViewContentMode.scaleAspectFit
-        
-        let url:URL = URL(string: urlString)!
-        let session = URLSession.shared
-        let task = session.dataTask(with: url) { (data, response, err) in
-            if data != nil{
-                let image = UIImage(data: data!)
-                if image != nil{
-                    DispatchQueue.main.async {
-                        imageView.image = image
-                    }
-                }
-            }else{
-                print("erreur dans la récupération de l'image")
-            }
-        }
-        task.resume()
-    }
-
+    
     func setCurrentPokemon(pokemon: Pokemon) {
         self.currentPokemon = pokemon
     }
